+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/AzureStorageClient.java
@@ -27,7 +27,7 @@ import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.util.zip.GZIPOutputStream;
 
-class AzureStorageClient {
+public class AzureStorageClient {
     private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     private static final int GZIP_BUFFER_SIZE = 16384;
     private static final int STREAM_BUFFER_SIZE = 16384;