+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/resources/QueueWithSas.java
@@ -1,15 +1,16 @@
-package com.microsoft.azure.kusto.ingest.utils;
+package com.microsoft.azure.kusto.ingest.resources;
 
 import com.azure.core.http.HttpClient;
 import com.azure.storage.common.policy.RequestRetryOptions;
 import com.azure.storage.queue.QueueClient;
 import com.azure.storage.queue.QueueClientBuilder;
+import com.azure.storage.queue.implementation.AzureQueueStorageImpl;
 import com.microsoft.azure.kusto.data.UriUtils;
 import reactor.util.annotation.Nullable;
 
 import java.net.URISyntaxException;
 
-public class QueueWithSas {
+public class QueueWithSas implements ResourceWithSas<QueueClient> {
     private final String sas;
     private final QueueClient queue;
 
@@ -34,5 +35,20 @@ public class QueueWithSas {
 
     public String getEndpoint() {
         return queue.getQueueUrl() + sas;
+    }
+
+    @Override
+    public String getEndpointWithoutSas() {
+        return queue.getQueueUrl();
+    }
+
+    @Override
+    public String getAccountName() {
+        return queue.getAccountName();
+    }
+
+    @Override
+    public QueueClient getResource() {
+        return queue;
     }
 }