+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/resources/RankedStorageAccount.java
@@ -1,0 +1,22 @@
+package com.microsoft.azure.kusto.ingest.resources;
+
+public class RankedStorageAccount {
+    private final String accountName;
+
+    public RankedStorageAccount(String accountName) {
+        this.accountName = accountName;
+    }
+
+    public void addResult(boolean success) {
+
+    }
+
+    public double getRank() {
+        return 1;
+    }
+
+    public String getAccountName() {
+        return accountName;
+    }
+
+}